+++ CHANGELOG.md
@@ -10,6 +10,7 @@
       - Add accelerated version to TransmissionAbsorption processor, controlled by `accelerated` parameter, default is True (#2036)
       - Made the call to next() in algorithm iteration loop explicit (#2069)
       - Added option for a random seed in the power method in the linear operator (#1585)
+      - Added option for `RANDOM_LOW_MEM` and `RANDOM_INT_LOW_MEM` arguments in geometry `allocate()` methods which use numpy default random number generator
   - Testing
       - Developers can now add `#all-tests` to their commit message on a PR to run the full matrix of GitHub actions tests (#2081)
   - Dependencies